+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Fixed
+- "Description" and "Usage" to README template
 - Badge link to package in README
 
 ## [1.2.0] - 2019-02-01