+++ CHANGELOG.md
@@ -5,8 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Fixed
+### Added
 - "Description" and "Usage" to README template
+
+### Fixed
+- Indenting in tsconfig.json template
 - Badge link to package in README
 
 ## [1.2.0] - 2019-02-01