+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.1.0] - 2019-02-22
 ### Added
 - Support for package scopes.
 
@@ -68,7 +68,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Script that deploys the templates into a project folder
 - Setup for this project
 
-[Unreleased]: https://github.com/Ionaru/create-package/compare/2.0.0...HEAD
+[Unreleased]: https://github.com/Ionaru/create-package/compare/2.1.0...HEAD
+[2.1.0]: https://github.com/Ionaru/create-package/compare/2.0.0...2.1.0
 [2.0.0]: https://github.com/Ionaru/create-package/compare/1.2.0...2.0.0
 [1.2.0]: https://github.com/Ionaru/create-package/compare/1.1.1...1.2.0
 [1.1.1]: https://github.com/Ionaru/create-package/compare/1.1.0...1.1.1